Pretty formatted code

+++ src/main/java/com/tripinfo/model/Trip.java
@@ -7,11 +7,11 @@ public class Trip {
     private final VehicleType vehicleType;
     private final double tripDistanceKm;
     private final double farePaid;
-    private final float customerRating;
+    private final double customerRating;
 
     // Full parameterised constructor
     public Trip(String driverId, VehicleType vehicleType,
-                double tripDistanceKm, double farePaid, float customerRating) {
+            double tripDistanceKm, double farePaid, Double customerRating) {
         this.driverId = driverId;
         this.vehicleType = vehicleType;
         this.tripDistanceKm = tripDistanceKm;
@@ -19,16 +19,32 @@ public class Trip {
         this.customerRating = customerRating;
     }
 
-    // We will not entertain an empty/default constructor, as that will embody an invalid trip.
+    // We will not entertain an empty/default constructor, as that will embody an
+    // invalid trip.
 
     // Parameter getter methods
-    public String getDriverId()        { return driverId; }
-    public VehicleType getVehicleType() { return vehicleType; }
-    public double getTripDistanceKm()  { return tripDistanceKm; }
-    public double getFarePaid()        { return farePaid; }
-    public float getCustomerRating()  { return customerRating; }
+    public String getDriverId() {
+        return driverId;
+    }
+
+    public VehicleType getVehicleType() {
+        return vehicleType;
+    }
+
+    public double getTripDistanceKm() {
+        return tripDistanceKm;
+    }
+
+    public double getFarePaid() {
+        return farePaid;
+    }
+
+    public double getCustomerRating() {
+        return customerRating;
+    }
+
     public double getEarningsPerKm() {
-        return (tripDistanceKm == 0.0) ? 0.0 : farePaid / tripDistanceKm; 
+        return (tripDistanceKm == 0.0) ? 0.0 : farePaid / tripDistanceKm;
         // Handled edge case where the trip distance is 0 (divide by zero error)
     }
 }
